move xtokenclient to transactors
