+++ index/src/transactors/mod.rs
@@ -1,2 +1,4 @@
 mod evm_transactor;
 pub use evm_transactor::ChainBridgeClient;
+mod xtokens;
+pub use xtokens::XtokenClient;
